* Several Bug fixes by Felipe Barriga Richards * all int's converted to long
BUG-2 Fixed

git-svn-id: https://svn.origo.ethz.ch/fcp2lib/trunk@11 f56a5a65-e1b4-4d35-b160-e61c0a6877f2

+++ FCP2lib/EventArgs/FCPPluginReplyEventArgs.cs
@@ -30,9 +30,9 @@ namespace Freenet.FCP2 {
             get { return pluginName; }
         }
         
-        private int? dataLength;
+        private long? dataLength;
         
-        public int? DataLength {
+        public long? DataLength {
             get { return dataLength; }
         }
         
@@ -61,7 +61,7 @@ namespace Freenet.FCP2 {
             
             this.pluginName = parsed["PluginName"];
             if(parsed["DataLength"] != null) {
-                this.dataLength = int.Parse(parsed["DataLength"]);
+                this.dataLength = long.Parse(parsed["DataLength"]);
                 
                 data = null; /* TODO: Similar to AllData*/
                 throw new NotImplementedException("Unclear format");
